Refactoring segmentation.
Code is now split into separate files (one per approach).
