+++ sima/segment/__init__.py
@@ -1,0 +1,7 @@
+from .segment import (
+    PlaneWiseSegmentationStrategy,
+    ROISizeFilter,
+    CircularityFilter)
+from .stica import PlaneSTICA
+from .ca1pc import PlaneCA1PC
+from .normcut import PlaneNormalizedCuts
