+++ sima/segment/_opca.pyx
