+++ sima/segment/tests/test_oPCA.py
